fix test case with use of SlFromRef and it's lack of mutable references
