--- builtins/trybuild/tests/borrow_checker_immut_mut_fail.stderr
@@ -1,9 +1,0 @@
-error[E0502]: cannot borrow `vm` as mutable because it is also borrowed as immutable
-  --> trybuild/tests/borrow_checker_immut_mut_fail.rs:11:37
-   |
-10 |     let y: &str = (&dest2).sl_as_ref(&vm).unwrap();
-   |                                      --- immutable borrow occurs here
-11 |     let x: String = (&dest).sl_into(&mut vm).unwrap();
-   |                                     ^^^^^^^ mutable borrow occurs here
-12 |     println!("y: {}", y);
-   |                       - immutable borrow later used here
